+++ user-service/src/main/java/sparta/userservice/user/UserRepository.java
@@ -6,7 +6,7 @@ import sparta.userservice.domain.User;
 import java.util.Optional;
 
 
-public interface UserRepository extends JpaRepository<User, Long> { // 엔티티 클래스, PK의 타입
+public interface UserRepository extends JpaRepository<User, Integer> { // 엔티티 클래스, PK의 타입
     Optional<User> findByName(String name);
     
     Optional<User> findByEmail(String email);
